Ban unused expressions
## [no-unused-expressions](https://eslint.org/docs/latest/rules/no-unused-expressions)
### Effect
Ban ternaries a shortcut for `if/else` and `&&` as a shortcut for `if` without assignment in TS files
### Purpose
With the team at the size it is and growing, readability should be prioritized, which means clear over compact and explicit over implicit

+++ front_end/libs/features/systems/settings/admin/standard/standard.component.ts
@@ -211,8 +211,9 @@ export class NxSystemStandardAdminComponent implements OnInit, OnChanges, OnDest
     }
 
     setValues(settings): void {
-        this.systemAndSecuritySettingsFormWatcher &&
+        if (this.systemAndSecuritySettingsFormWatcher) {
             this.applyService.removeFormWatcher('systemAndSecuritySettingsForm');
+        }
         const sw = this.systemAndSecuritySettings;
         const useRest = this.system.useRest;
         Object.keys(sw).forEach(setting => {
